[Rebase] create Main Screen
