+++ lib/ui/common/app_page.dart
@@ -1,14 +1,18 @@
 import 'package:get/get.dart';
+import 'package:todo_app/ui/main/main_view.dart';
+
+import '../../router/route_name.dart';
+import '../main/main_binding.dart';
 
 class AppPage {
 
   static final List<GetPage> pages = [
 
-    // GetPage(
-    //   name: AppRouteName.homeTab,
-    //   page: () => const HomeTabView(),
-    //   binding: HomeTabBinding(),
-    // ),
+    GetPage(
+      name: AppRouteName.main,
+      page: () => const MainView(),
+      binding: MainBinding(),
+    ),
 
   ];
